refactor: 516117 - Create new migration service
